+++ extension/firefox/src/background/background.ts
@@ -1,1 +1,3 @@
-export {};
+import { installFirefoxDocumentRedirect } from './firefox-redirect';
+
+installFirefoxDocumentRedirect();